webuiapi/pyproject.toml
[project]
name = "webuiapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "comfy-api-simplified>=1.4.2",
    "pillow>=12.1.0",
    "websockets>=15.0.1",
]
